Introducing FileContext and Projects
