+++ app/src/main/java/com/kyhsgeekcode/disassembler/ILAssmebly.java
@@ -14,16 +14,6 @@ public class ILAssmebly extends AbstractFile {
     }
 
     @Override
-    public void setPath(String path) {
-        super.setPath(path);
-    }
-
-    @Override
-    public String getPath() {
-        return super.getPath();
-    }
-
-    @Override
     public MachineType getMachineType() {
         return super.getMachineType();
     }
