+++ app/src/main/java/com/kyhsgeekcode/disassembler/RawFile.java
@@ -5,7 +5,7 @@ import java.io.File;
 public class RawFile extends AbstractFile {
     public RawFile(File file, byte[] filecontent) {
         this.fileContents = filecontent;
-        setPath(file.getPath());
+        setPath(file.getAbsolutePath());
 
     }
